+++ src/mvvm/value_converter.h
@@ -76,6 +76,49 @@ ret_t value_converter_to_model(value_converter_t* converter, const value_t* from
  */
 ret_t value_converter_to_view(value_converter_t* converter, const value_t* from, value_t* to);
 
+/**
+ * @method value_converter_create
+ * 创建指定名称的值转换器。
+ *
+ * @annotation ["scriptable", "static"]
+ * @param {const char*} name 名称。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+value_converter_t* value_converter_create(const char* name);
+
+/**
+ * @method value_converter_register
+ *
+ * 注册值转换器的创建函数。
+ * @annotation ["scriptable", "static"]
+ * @param {const char*} name 名称。
+ * @param {tk_create_t} create 创建函数。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t value_converter_register(const char* name, tk_create_t create);
+
+/**
+ * @method value_converter_init
+ *
+ * 初始化值转换器的全局对象。
+ * @annotation ["scriptable", "static"]
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t value_converter_init(void);
+
+/**
+ * @method value_converter_deinit
+ *
+ * 释放值转换器的全局对象。
+ * @annotation ["scriptable", "static"]
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t value_converter_deinit(void);
+
 #define VALUE_CONVERTER(converter) ((value_converter_t*)(converter))
 
 END_C_DECLS