+++ src/tkc/object.c
@@ -111,6 +111,15 @@ const char* object_get_prop_str(object_t* obj, const char* name) {
   }
 }
 
+void* object_get_prop_pointer(object_t* obj, const char* name) {
+  value_t v;
+  if (object_get_prop(obj, name, &v) == RET_OK) {
+    return value_pointer(&v);
+  } else {
+    return NULL;
+  }
+}
+
 int32_t object_get_prop_int(object_t* obj, const char* name, int32_t defval) {
   value_t v;
   if (object_get_prop(obj, name, &v) == RET_OK) {
@@ -156,6 +165,13 @@ ret_t object_set_prop(object_t* obj, const char* name, const value_t* v) {
 ret_t object_set_prop_str(object_t* obj, const char* name, const char* value) {
   value_t v;
   value_set_str(&v, value);
+
+  return object_set_prop(obj, name, &v);
+}
+
+ret_t object_set_prop_pointer(object_t* obj, const char* name, void* value) {
+  value_t v;
+  value_set_pointer(&v, value);
 
   return object_set_prop(obj, name, &v);
 }